Move pinch to zoom to new file
