+++ Package.swift
@@ -27,9 +27,9 @@ let package = Package(
         // Targets can depend on other targets in this package, and on products in packages which this package depends on.
         .target(
             name: "ImageViewer",
-            dependencies: ["URLImage"]),
+            dependencies: []),
         .target(
             name: "ImageViewerRemote",
-            dependencies: ["URLImage"])
+            dependencies: ["ImageViewer", "URLImage"])
     ]
 )
